requirements.txt
watchdog
openai-whisper
moviepy
ffmpeg-python
numpy
torch
